requirements.txt
pytorch
scikit-learn
openai
streamlit
opencv-python
pillow
